requirements.txt
Django
apscheduler

django-apscheduler
django-widget-tweaks
django-storages
django-redis
django-cache-memoize
djangorestframework
djangorestframework_simplejwt
djangorestframework-datatables
django-filter
django-cors-headers
django-rest-auth
drf-extensions
django-sslserver
django_prometheus

# Database encrypton
# django-fernet-fields  # Not working with DJango >= 4
djfernet

# Nomad + Consul clients
python-nomad
python-consul

# Vault SDK
hvac[parser]

# Docker client
docker
pyyaml

# Utility packages
requests
simplejson

# Database connectors
mysqlclient
psycopg2

redis
pika

boto3